bot/requirements.txt
requests~=2.28.1
python-telegram-bot